Provide an helper method that can be used to parse both version of OPDS feeds
